backend/requirements.txt
Django>=3.2
psycopg2-binary
python-decouple
